Reorganize page and component file structure; Separate Hero and Navbar from the Homepage
