+++ react-frontend/src/App.js
@@ -1,4 +1,4 @@
-import Homepage from './components/Homepage';
+import Homepage from './pages/Homepage';
 
 function App() {
   return (
